requirements.txt
absl-py==1.2.0
antlr4-python3-runtime==4.9.3
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
astroid==2.11.7
async-generator==1.10
atari-py==0.2.9
attrs==21.4.0
Automat==20.2.0
autopep8==1.6.0
Babel==2.10.3
backcall==0.2.0
bcrypt==3.2.2
bleach==4.1.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2020.6.20
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==1.2.2
colorama==0.4.5
constantly==15.1.0
contextvars==2.4
cryptography==37.0.4
cycler==0.11.0
Cython==0.29.30
-e git+https://github.com/rail-berkeley/d4rl.git@4aff6f8c46f62f9a57f79caa9287efefa45b6688#egg=d4rl
dataclasses==0.8
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.4
docopt==0.6.2
dotmap==1.3.30
entrypoints==0.4
fasteners==0.17.3
flake8==4.0.1
future==0.18.2
gast==0.3.3
gitdb==4.0.9
GitPython==3.1.18
glfw==2.5.4
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
gtimer==1.0.0b5
gym==0.15.7
h5py==3.1.0
hope==3.6.0
hyperlink==21.0.0
idna==3.3
imageio==2.15.0
immutables==0.18
importlib-metadata==4.8.3
importlib-resources==5.4.0
incremental==21.3.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.2
isort==5.10.1
jedi==0.17.2
jedi-language-server==0.21.0
Jinja2==3.0.3
joblib==1.1.0
json5==0.9.8
jsonschema==4.0.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyter-lsp==1.5.1
jupyter-server==1.13.1
jupyter-server-mathjax==0.2.3
jupyterlab==3.2.9
jupyterlab-code-cell-collapser==1.0.0
jupyterlab-git==0.37.1
jupyterlab-lsp==3.9.2
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.1.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.7.1
lockfile==0.12.2
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.3.4
mccabe==0.6.1
mistune==0.8.4
mujoco-py==2.0.2.13
nbclassic==0.3.5
nbclient==0.5.9
nbconvert==6.0.7
nbdime==3.1.1
nbformat==5.1.3
nest-asyncio==1.5.5
notebook==6.4.10
numpy==1.19.5
oauthlib==3.2.0
omegaconf==2.3.0
opencv-python==4.6.0.66
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
paramiko==2.11.0
parso==0.7.1
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pipreqs==0.4.11
platformdirs==2.4.0
pluggy==1.0.0
prettytable==2.5.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
PuLP==2.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybullet==3.2.5
pycodestyle==2.8.0
pycparser==2.21
pydocstyle==6.1.1
pyflakes==2.4.0
pyglet==1.5.0
pygls==0.9.1
Pygments==2.12.0
pylint==2.13.9
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.18.0
pysftp==0.2.9
python-dateutil==2.8.2
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.3.3
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.2.2
requests==2.27.1
requests-oauthlib==1.3.1
-e git+ssh://git@github.com/polixir/RLAssistant.git@4efee6599481c57182bf1f5ae1643ad58a8f52bd#egg=RLA
rope==1.1.1
rsa==4.9
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
Send2Trash==1.8.0
Serializable @ git+https://github.com/hartikainen/serializable.git@76516385a3a716ed4a2a9ad877e2d5cbcf18d4e6
service-identity==21.1.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
sniffio==1.2.0
snowballstemmer==2.2.0
sshim==1.3.3
stable-baselines==2.10.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.13.0
testpath==0.6.0
threadpoolctl==3.1.0
toml==0.10.2
tomli==1.2.3
torch==1.3.0
tornado==6.1
tqdm==4.64.0
traitlets==4.3.3
Twisted==22.4.0
typed-ast==1.5.4
typing==3.7.4.3
typing_extensions==4.1.1
ujson==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
Werkzeug==2.0.3
widgetsnbextension==3.6.1
wrapt==1.14.1
yapf==0.32.0
yarg==0.1.9
zipp==3.6.0
zope.interface==5.4.0
